Update main.py
Updates the search colors and how the boxes work. Now there can be up to 10 boxes for a single color by using decimal rounding, this means that it will round the number that it told to look for so 2.1 will show up as 2 but 2.1 could have different settings than 2. This gives the program more versatility and options. In the end when you search for the number in the ObjectLocator 2.1, 1.9... will all become a 2 because they were rounded.